+++ src/detectors/__init__.py
@@ -1,3 +1,4 @@
 from .apd import apd
+from .spad import spad
 
-__all__ = ['apd']
+__all__ = ['apd', 'spad']